refactor: changed app struct
