+++ src/app/modules/chanels_module/chanel-card/channel-card.component.html
@@ -6,6 +6,7 @@
 
 <mat-card class="channel_card">
   <img src="{{info.logo}}" alt="{{info.name + 'channel logo'}}">
-  <p>{{info.name}}</p>
-  <p>{{info.age_rating}}</p>
+  <span>{{info.name}}</span>
+  <span>{{info.age_rating}}</span>
+  <span>Rank: {{info.rank}}</span>
 </mat-card>
