+++ src/app/modules/chanels_module/chanel-card/channel-card.component.ts
