+++ src/app/modules/chanels_module/channels/channels.component.html
@@ -1,8 +1,6 @@
 <div class="wrapper">
   <div class="channels">
-    <p>
-      chanells
-    </p>
+    <h1>Channels</h1>
 
     <div *ngFor="let obj of test">
       <app-channel-card [info]="obj"></app-channel-card>
