+++ src/app/modules/chanels_module/channels/channels.component.scss
@@ -1,7 +1,3 @@
-.wrapper {
-  width: 950px;
-}
-
 .channels {
   display: flex;
   align-items: center;
