+++ src/app/modules/chanels_module/channels/channels.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from "@angular/core";
 import {HttpClient} from "@angular/common/http";
-import {GetchannelsService} from "../services/getchannels.service";
+import {GetchannelsService} from "../../../services/getchannels.service";
 import {ChannelModel} from "../interfaces/channel.model";
