+++ src/app/modules/chanels_module/interfaces/channel.model.ts
